requirements.txt
datasets=3.2.0
huggingface-hub=0.26.2
pandas=2.2.3
pillow=10.3.0
pyarrow=18.0.0
pytorch-cuda=12.1
scikit-learn=1.5.2
scipy=1.14.1
sentencepiece=0.2.0
setuptools=75.3.0
tokenizers=0.20.1
torchaudio=2.5.1
torchtriton=3.1.0
torchvision=0.20.1
tqdm=4.66.6
transformers=4.46.1